// src/app/pages/model/episode.ts
export class Episode {
    baby_id: any='';
    admit_to: number=null;
    local_hospital_id='';
    admit_time ='';
    weight: number = null;
    head_circumference: number = null;
    temperature: number = null;
    admit_from:number=null;
    admit_from_other: string;
    type='';
    id: any;
}